17 addd support for actuator (#25)
* fix: update Makefile with environment

This commit introduces a separate development environment for the application, enabling independent deployments and management of development and production instances.  The Makefile, deployment scripts, and configuration files have been updated to support this change.  A new URL for accessing the Swagger UI in the development environment is also provided.

* enable actuator info endpoint

trigger build